x86: make spinlock's 16-bit asm operand explicitly 16-bit. This is needed to compile xen with clang.
Acked-by: Keir Fraser <keir@xen.org>

+++ xen/include/asm-x86/spinlock.h
@@ -27,7 +27,7 @@ static always_inline int _raw_spin_trylock(raw_spinlock_t *lock)
     asm volatile (
         "xchgw %w0,%1"
         :"=r" (oldval), "=m" (lock->lock)
-        :"0" (0) : "memory" );
+        :"0" ((s16)0) : "memory" );
     return (oldval > 0);
 }
